Move evaluators/ and model_interface/ out of common/
