+++ loss_landscapes/__init__.py
@@ -3,4 +3,4 @@ from loss_landscapes.main import linear_interpolation
 from loss_landscapes.main import random_line
 from loss_landscapes.main import planar_interpolation
 from loss_landscapes.main import random_plane
-from loss_landscapes.common.model_interface.agent_interface import AgentInterface
+from loss_landscapes.model_interface.agent_interface import AgentInterface
