+++ loss_landscapes/evaluators/torch/sl_evaluators.py
@@ -11,10 +11,10 @@ from abc import ABC, abstractmethod
 import numpy as np
 import torch
 import torch.autograd
-from loss_landscapes.common.evaluators.evaluators import Evaluator
-from loss_landscapes.common.model_interface.agent_interface import wrap_model
-from loss_landscapes.common.model_interface.tensor_factory import rand_u_like
-from loss_landscapes.common.model_interface.torch.torch_wrappers import TorchModelWrapper
+from loss_landscapes.evaluators.evaluators import Evaluator
+from loss_landscapes.model_interface.agent_interface import wrap_model
+from loss_landscapes.model_interface.tensor_factory import rand_u_like
+from loss_landscapes.model_interface.torch.torch_wrappers import TorchModelWrapper
 
 
 class TorchSupervisedEvaluator(Evaluator, ABC):
