+++ loss_landscapes/model_interface/torch/torch_wrappers.py
@@ -3,8 +3,8 @@ Defines functions for getting and setting the parameters of a model.
 """
 
 import copy
-import loss_landscapes.common.model_interface.model_wrapper as model_wrapper
-import loss_landscapes.common.model_interface.torch.torch_tensor as torch_tensor
+import loss_landscapes.model_interface.model_wrapper as model_wrapper
+import loss_landscapes.model_interface.torch.torch_tensor as torch_tensor
 
 
 class TorchModelWrapper(model_wrapper.ModelWrapper):
