+++ loss_landscapes/trajectories.py
@@ -7,7 +7,7 @@ a low-dimensional approximation of the trajectory.
 from abc import ABC, abstractmethod
 from datetime import datetime
 import numpy as np
-from loss_landscapes.common.model_interface.agent_interface import wrap_model
+from loss_landscapes.model_interface.agent_interface import wrap_model
 
 
 class TrajectoryTracker(ABC):
